+++ packages/router/src/lib/errors/fulfill.ts
@@ -2,7 +2,7 @@ import { NxtpError } from "@connext/nxtp-utils";
 
 export class NoChainConfig extends NxtpError {
   constructor(chainId: number, context: any = {}) {
-    super(`No chain config for chainId ${chainId}`, context, "NoChainConfig");
+    super(`No chain config for chainId ${chainId}`, context, "NoChainConfig", "debug");
   }
 }
 
@@ -14,6 +14,6 @@ export class NotEnoughRelayerFee extends NxtpError {
 
 export class NotAllowedFulfillRelay extends NxtpError {
   constructor(chainId: number, context: any = {}) {
-    super(`Not allowed fulfill relay, ${chainId}`, context, "NotAllowFulfillRelay");
+    super(`Not allowed fulfill relay, ${chainId}`, context, "NotAllowFulfillRelay", "debug");
   }
 }