+++ packages/router/src/lib/operations/auction.ts
@@ -155,7 +155,7 @@ export const newAuction = async (
   }
 
   // Make sure subgraphs are synced
-  const receivingSyncRecords = await contractReader.getSyncRecord(receivingChainId, requestContext);
+  const receivingSyncRecords = await contractReader.getSyncRecords(receivingChainId, requestContext);
   if (!receivingSyncRecords.some((record) => record.synced)) {
     throw new SubgraphNotSynced(receivingChainId, receivingSyncRecords, {
       methodContext,
@@ -164,7 +164,7 @@ export const newAuction = async (
     });
   }
 
-  const sendingSyncRecords = await contractReader.getSyncRecord(sendingChainId, requestContext);
+  const sendingSyncRecords = await contractReader.getSyncRecords(sendingChainId, requestContext);
   if (!sendingSyncRecords.some((record) => record.synced)) {
     throw new SubgraphNotSynced(sendingChainId, sendingSyncRecords, {
       methodContext,